api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.28.1
kafka-python==2.0.2
